updated migrations

+++ database/migrations/2021_04_29_103409_create_products_table.php
@@ -6,7 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateProductsTable extends Migration
 {
-    
     /**
      * Run the migrations.
      *
@@ -16,8 +15,11 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('seller_id');
-            $table->foreign('seller_id')->references('id')->on('sellers');
+            $table->unsignedBigInteger('seller_id');
+            $table
+                ->foreign('seller_id')
+                ->references('id')
+                ->on('sellers');
             $table->string('type');
             $table->text('desc');
             $table->double('price');
