+++ database/migrations/2021_04_29_103611_create_orders_table.php
@@ -15,10 +15,16 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('product_id');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('product_id')->references('id')->on('products');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
+            $table
+                ->foreign('user_id')
+                ->references('id')
+                ->on('users');
+            $table
+                ->foreign('product_id')
+                ->references('id')
+                ->on('products');
             $table->integer('qty');
             $table->double('price');
             $table->double('discount');
